hook: limit subject to 72 characters
The limit of 50 characters is a statistical average of what has been
found as a good commit subject.  While that should be the goal, it is
not the limit for the length of the subject line and often forces
incorrect messages to be used.  The subject line should be short, as
short as possible, without be obscure.

Signed-off-by: Dan Kalowsky <dank@deadmime.org>